Fix incorrect use of scoped enumerations in format strings
Scoped enums do not get automatically promoted to their underlying type,
so these uses have undefined behavior and Clang recently started warning
about it.

Bug: 1456289
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4632727
Commit-Queue: Nico Weber <thakis@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Owners-Override: Nico Weber <thakis@chromium.org>
Auto-Submit: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161189}

+++ ios/web/navigation/crw_wk_navigation_states.mm
@@ -56,7 +56,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef NDEBUG
 - (NSString*)description {
   return [NSString stringWithFormat:@"state: %d, index: %ld, context: %@",
-                                    _state, static_cast<long>(_index),
+                                    static_cast<int>(_state),
+                                    static_cast<long>(_index),
                                     _context->GetDescription()];
 }
 #endif  // NDEBUG
